requirements.txt
pytest
playwright
pytest-playwright
allure-pytest
pytest-xdist
python-dotenv